requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
python-dotenv==1.0.1
sqlmodel==0.0.21
aiosqlite==0.20.0
uvloop==0.20.0; platform_system != "Windows"
orjson==3.10.7
python-multipart==0.0.12

